Fix output image visualisation in image preprocessing examples (#199)
* Extract ImagePanel class to use in image preprocessing examples

* Display images in the ImagePanel as is

Convert the array to a BufferedImage. Do not display black borders around each pixel and keep the color as is.

* Fix image frame size

Pass final shape of the Preprocessing object to the ImagePanel. Calculate preferred size of the panel from the image size. Use Window#pack to resize the frame according to the subcomponents.