Move loading out of preprocessing (#323)
* Require ImageShape to be not null for getOutputShape method

As ImageShape can have null values for width, height and channels, passing null to indicate that input dimensions are unknown is not necessary.

* Remove imageShape property from preprocessing loading stage

The input shape is used only when finalShape for the preprocessing is calculated, so it could be passed as a parameter when needed.

* Move loading out of preprocessing

Loading functionality seems to be orthogonal to the preprocessing functionality, and fits much better in the Dataset implementations, as the knowledge of how images are loaded is intrinsic to them. Without the loading stage, preprocessors could be shared between datasets and applied to the images loaded from various sources.