Add new preprocessing DSL (#416)
New preprocessing API implies that preprocessing operations are defined as functional
objects with Input and Output types defined as generics.
Operations can be composed together to form a pipeline. The output of the operation
is the input of the next operation. At the end of the pipeline, the output of the
last operation is the pipeline's output.

The approach when the pipeline is defined as a sequence of operations has the
following advantages:
 - it's more flexible and supports the composition of operations with different
   input and output types.
 - user explicitly specifies conversions between types during the pipeline definition. E.g. there is an Operation that converts from [BufferedImage] to [FloatArray].
 - allows adding support of augmentations in the future without changing the API.

* Remove imagePreprocessing and tensorPreprocessing DSL blocks

* User explicitly convert BufferedImage to FloatArray

* Replace ImageShape with TensorShape

* getOutputShape method now implemented for tensor operations
  and now calculates correct shape for operations such as Transpose

* Update tests and examples with new API

Co-authored-by: Julia Beliaeva <Julia.Beliaeva@jetbrains.com>