Refactor swing visualization (#451)
* Remove unused code

* Move detected result visualization to a separate file

* Remove visualization methods for the raw output

Always convert raw output to the high-level objects. This allows to reuse the same visualization methods and to get rid of the extra code.

* Extract method to show a frame with content

* Move ImagePanel to the visualization package

* Use BufferedImage to visualize detection results

This allows to get rid of the unnecessary conversion of the image to FloatArray and back, and use the input image as is (applying preprocessing only to resize it when needed).

* Extract functions to draw detection results on a graphics and remove duplication

* Enhance detection results visualization

  * Use antialiasing
  * Use float coordinates instead of integers
  * Draw pose edges under vertices
  * Use default font face and size for labels
  * Adjust labels location
  * Specify precision for outputting probability
  * Cleanup code

* Replace functions showing frames with functions returning panels

This allows to show detection results for several images in one frame.

* Skip low-confidence pose detection results

* Add documentation