Remove ImageShape class (#491)
After refactoring of the preprocessing dsl in 5a5d385fb56fb8d9d44f63430fffd3c3bf7faf34, the original function of ImageShape is no longer there as the new Operation interface uses TensorShape. ImageShape is only used in the floatArrayToBufferedImage functions and in tests, and can be easily replaced by TensorShape or just a pair of width and height.