Cleanup model copying (#503)
* Remove extra parameters from the InferenceModel#copy

copyWeight and saveOptimizerState parameters are unused in the OnnxInferenceModel, and InferenceModel is not required to have a name, so copiedModelName does not always make sense as well. It's better to have a more generic copy method in the interface, and move these parameters to the appropriate implementing classes.

* Remove extra parameters from the TensorFlowInferenceModel#copy

TensorFlowInferenceModel does not have any layers, so it only makes sense to copy weights. And since the model class is not trainable, optimizer state does not need to be copied.

* Support copying optimizer variables in Sequential and Functional models