Upgraded the lets-plot library and Kotlin API to the 4.x version and migrated the API (#569)
* Update Lets-Plot library and adjust import paths

Updated the Lets-Plot library in build.gradle from version 2.3.0 to 4.0.1 and 2.0.1 to 4.4.3.

* Refined comments and modified code for clarity

Refactored code for better readability. Corrected grammar and improved sentence structure in comments for better understanding. Also, tweaked theme configuration in 'PlotDefinition.kt' for explicit settings. Updated example comment on 'SoundNetFSDDVisualization.kt' to be singular as it demonstrates a single example.